+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - feat: CodeVm - add authorization via OTP
 - feat: SignInScreenVm - add phone authorization
 - feat: OnboardingScreen
+- feat: MainScreen + PlacesScreen
 
 ## 0.3.0